// web/src/types/moneyFlow.ts
export type MoneyFlowMetrics = {
  workingBudget: number;
  savingsGoal: number;
  spendingLimit: number;
  monthSpent: number;
  remainingSpending: number;
  deficit: number;
  reserveTotal: number;
  savingsShare: number;
  spendingShare: number;
  spentShare: number;
  remainingShare: number;
  hasBudget: boolean;
  isOverBudget: boolean;
};